Testbed stage: extract MemtermCore library and add behavioral test suite
- New MemtermCore target holding the pure logic, AppKit/SwiftTerm-free so
  it tests headlessly: TOML subset parser + Config (colors as ConfigRGB
  triples, font resolution split into pure resolveFontName), StateStore +
  split-tree encode/decode + restore models, Adapters (classification,
  denylist, shellQuote, claude session-file resolution), CwdFallback
  (FR-25), ScrollbackText assembly. Executable maps ConfigRGB -> NSColor /
  SwiftTerm.Color and font names -> NSFont in ConfigUI.swift.
- Injectable file locations for tests: StateStore takes its db URL,
  Config.load(from:), Adapters.classify/claudeSessionId take a projects
  dir; defaults unchanged (App Support / ~/.config / ~/.claude).
- Denylist hardened per FR-30: curl/wget piped into a shell is now
  rejected (was a gap); CwdFallback now lands on $HOME instead of "/"
  when a saved cwd is fully vanished (FR-25 "else $HOME").
- MemtermCoreTests: 55 behavioral tests — config parsing (defaults file,
  escapes, inline comments, bad/unknown keys, font candidate order),
  StateStore (topology round-trip, WAL reopen + second connection,
  no-orphan rewrites, torn-write corruption = degraded open never crash,
  concurrent write serialization), adapters (argv fixtures, denylist,
  shellQuote round-trip through /bin/sh, claude cwd->slug), cwd fallback,
  split-tree JSON round-trip, scrollback assembly.
- Gates re-verified: release build green, swift test 55/55, --smoke pair
  green (1/2/3 saved and restored), --bench/--latency/--flood green.
  Consent invariants untouched: ⌘R still types without newline, denylist
  precedes offers, ghost scrollback is feed()-only.
- Also carries founder edits pending in the worktree: REQUIREMENTS
  FR-56..58 (deliberate-close-forgets, kill memories, right-click
  affordances) and the app icon assets.

+++ Sources/memterm/MemtermApp.swift
@@ -1,4 +1,5 @@
 import AppKit
+import MemtermCore
 
 // Interactive-mode application delegate: config, main menu, window/tab/pane
 // lifecycle. The probe modes (--latency/--flood) keep their own bare delegate
@@ -15,7 +16,7 @@ final class MemtermAppDelegate: NSObject, NSApplicationDelegate {
 
     init(smokeMode: Bool = false) {
         self.smokeMode = smokeMode
-        fontSize = config.fontSize
+        fontSize = CGFloat(config.fontSize)
         super.init()
     }
